% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_latex.R
\name{tex_hline}
\alias{tex_hline}
\title{Return a booktabs horizontal rule}
\usage{
tex_hline()
}
\value{
\code{character(1)} — \code{"\\midrule"}.
}
\description{
Return a booktabs horizontal rule
}
\examples{
tex_hline()

}
