% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{to_excel}
\alias{to_excel}
\title{Export a regression table to Excel}
\usage{
to_excel(x, file, sheet = "Table 1", open = FALSE)
}
\arguments{
\item{x}{A \code{regtab_table} object.}

\item{file}{Character. Output file path (required). Should end in \code{.xlsx}.}

\item{sheet}{Character. Worksheet name. Default \code{"Table 1"}.}

\item{open}{Logical. If \code{TRUE}, attempt to open the file after writing.
Default \code{FALSE}.}
}
\value{
\code{x} invisibly.
}
\description{
Writes a formatted Excel workbook using the \code{openxlsx2} package.
No Java dependency.
}
\examples{
\dontrun{
m1 <- lm(mpg ~ wt + cyl, data = mtcars)
tab <- regtab(list("(1)" = m1))
to_excel(tab, file = "table1.xlsx")
}

}
